+++ CHANGELOG.md
@@ -3,6 +3,7 @@
 ## Unreleased
 
 - Add `traces_sampler` option to set custom sample rate callback (#1083)
+- [BC BREAK] Add named constructors to the `Event` class (#1085)
 
 ## 3.0.0-beta1 (2020-09-03)